+++ src/Twilio/Rest/Preview/Understand/Assistant/Intent/FieldOptions.cs
@@ -115,7 +115,7 @@ namespace Twilio.Rest.Preview.Understand.Assistant.Intent
         /// </summary>
         public string PathIntentSid { get; }
         /// <summary>
-        /// The unique name or sid of the FieldType
+        /// The unique name or sid of the FieldType. It can be any Built-in Field Type or the unique_name or sid of a custom Field Type.
         /// </summary>
         public string FieldType { get; }
         /// <summary>
@@ -128,7 +128,8 @@ namespace Twilio.Rest.Preview.Understand.Assistant.Intent
         /// </summary>
         /// <param name="pathAssistantSid"> The assistant_sid </param>
         /// <param name="pathIntentSid"> The intent_sid </param>
-        /// <param name="fieldType"> The unique name or sid of the FieldType </param>
+        /// <param name="fieldType"> The unique name or sid of the FieldType. It can be any Built-in Field Type or the
+        ///                 unique_name or sid of a custom Field Type. </param>
         /// <param name="uniqueName"> A user-provided string that uniquely identifies this resource as an alternative to the
         ///                  sid. Unique up to 64 characters long. </param>
         public CreateFieldOptions(string pathAssistantSid, string pathIntentSid, string fieldType, string uniqueName)