+++ CHANGES.md
@@ -1,6 +1,16 @@
 twilio-csharp Changelog
 =======================
 
+[2018-07-05] Version 5.13.7
+----------------------------
+**Library**
+- PR #427: Add Dockerfile and related changes to build the Docker image. Thanks to @jonatasbaldin!
+
+**Api**
+- Release `Call Recording Controls` feature support in helper libraries
+- Add Voice Insights sub-category keys to usage records
+
+
 [2018-06-29] Version 5.13.6
 ----------------------------
 **Library**